+++ example/hello_error.ml
@@ -26,7 +26,7 @@
 let%expect_test "hello error" =
   Eio_main.run
   @@ fun env ->
-  let vcs = Vcs_git.create ~env in
+  let vcs = Vcs_git_eio.create ~env in
   let invalid_path = Absolute_path.v "/invalid/path" in
   let redact_sexp sexp =
     (* Because the actual error may become too brittle overtime, we actually