+++ example/hello_git_cli.ml
@@ -28,7 +28,7 @@ let%expect_test "hello cli" =
   @@ fun env ->
   Eio.Switch.run
   @@ fun sw ->
-  let vcs = Vcs_git.create ~env in
+  let vcs = Vcs_git_eio.create ~env in
   let repo_root = Vcs_test_helpers.init_temp_repo ~env ~sw ~vcs in
   let hello_file = Vcs.Path_in_repo.v "hello.txt" in
   Vcs.save_file