+++ lib/vcs_git_eio/test/test__hello_commit.ml
@@ -27,7 +27,7 @@ let%expect_test "hello commit" =
   @@ fun env ->
   Eio.Switch.run
   @@ fun sw ->
-  let vcs = Vcs_git.create ~env in
+  let vcs = Vcs_git_eio.create ~env in
   let mock_revs = Vcs.Mock_revs.create () in
   let repo_root = Vcs_test_helpers.init_temp_repo ~env ~sw ~vcs in
   let hello_file = Vcs.Path_in_repo.v "hello.txt" in